#92: refactor lawm fortran tests
